[ TEST ] fix compilation error
This commit includes,
  . compilation error fix : unused + wrong type comparison

**Self evaluation:**
1. Build test:	 [X]Passed [ ]Failed [ ]Skipped
2. Run test:	 [X]Passed [ ]Failed [ ]Skipped

+++ Applications/Custom/LayerPlugin/layer_plugin_common_test.cpp
@@ -58,8 +58,8 @@ TEST_P(LayerPluginCommonTest, DefaultEnvironmentPath_p) {
   }
 
   EXPECT_THROW(lnode->setProperty({"invalid_values"}), std::invalid_argument);
-  EXPECT_EQ(lnode->getOutputDimensions().size(), 0);
-  EXPECT_EQ(lnode->getInputDimensions().size(), 0);
+  EXPECT_EQ(lnode->getOutputDimensions().size(), size_t(0));
+  EXPECT_EQ(lnode->getInputDimensions().size(), size_t(0));
 }
 
 TEST_P(LayerPluginCommonTest, DefaultEnvironmentPathLayerNotExist_n) {
